+++ src/lib/rxfm/utils/utils.ts
@@ -1,5 +1,5 @@
-import { combineLatest, Observable, of, OperatorFunction } from "rxjs";
-import { distinctUntilChanged, map, pluck, shareReplay, switchMap, tap } from "rxjs/operators";
+import { combineLatest, MonoTypeOperatorFunction, Observable, of, OperatorFunction } from "rxjs";
+import { distinctUntilChanged, ignoreElements, map, pluck, shareReplay, startWith, switchMap, tap } from "rxjs/operators";
 
 export function coerceToObservable<T>(value: T | Observable<T>): Observable<T> {
   return value instanceof Observable ? value : of(value);
@@ -186,7 +186,7 @@ export function reuse<T>(source: Observable<T>): Observable<T> {
 /**
  * @returns An observable emitting the logical NOT value of the source observable's emissions.
  */
- export function notGate(source: Observable<any>): Observable<boolean> {
+ export function not(source: Observable<any>): Observable<boolean> {
   return source.pipe(
     distinctUntilChanged(),
     map(val => !val),
@@ -197,7 +197,7 @@ export function reuse<T>(source: Observable<T>): Observable<T> {
 /**
  * Take a spread array of observables and emit the logical AND value of all of their emissions whenever it changes.
  */
-export function andGate(...sources: Observable<any>[]): Observable<boolean> {
+export function and(...sources: Observable<any>[]): Observable<boolean> {
   return combineLatest(
     sources.map(source => source.pipe(distinctUntilChanged())),
   ).pipe(
@@ -209,14 +209,14 @@ export function andGate(...sources: Observable<any>[]): Observable<boolean> {
 /**
  * Take a spread array of observables and emit the logical NAND value of all of their emissions whenever it changes.
  */
-export function nandGate(...sources: Observable<any>[]): Observable<boolean> {
-  return notGate(andGate(...sources));
+export function nand(...sources: Observable<any>[]): Observable<boolean> {
+  return not(and(...sources));
 }
 
 /**
  * Take a spread array of observables and emit the logical OR value of all of their emissions whenever it changes.
  */
-export function orGate(...sources: Observable<any>[]): Observable<boolean> {
+export function or(...sources: Observable<any>[]): Observable<boolean> {
   return combineLatest(
     sources.map(source => source.pipe(distinctUntilChanged())),
   ).pipe(
@@ -228,8 +228,8 @@ export function orGate(...sources: Observable<any>[]): Observable<boolean> {
 /**
  * Take a spread array of observables and emit the logical NOR value of all of their emissions whenever it changes.
  */
-export function norGate(...sources: Observable<any>[]): Observable<boolean> {
-  return notGate(orGate(...sources));
+export function nor(...sources: Observable<any>[]): Observable<boolean> {
+  return not(or(...sources));
 }
 
 /**
@@ -258,5 +258,24 @@ export function log<T = unknown>(message?: string | ((val: T) => string)): Opera
       else if (typeof message === 'string') console.log(message, val);
       else if (typeof message === 'function') console.log(message(val));
     }),
+  );
+}
+
+/**
+ * A function taking an observable stream of type T and adding a side effect observable into the stream.
+ * This is similar to the built in `tap` operator in RxJS except the side effect is an observable to be injected into the stream.
+ * @param effect A function taking the emission of the source observable and returning an observable of any type.
+ * @returns An observable mirroring the source observable.
+ */
+ export function switchTap<T, U>(
+  effectObservable: (element: T) => Observable<U>,
+): MonoTypeOperatorFunction<T> {
+  return (source: Observable<T>) => source.pipe(
+    distinctUntilChanged(),
+    switchMap(element => effectObservable(element).pipe( // Add the effect observable into the stream.
+      ignoreElements(), // Ignore the effect observable's emissions.
+      startWith(element), // Return the original element as a single emission.
+    )),
+    distinctUntilChanged(),
   );
 }